Fixed binder tests.
